+++ app/src/main/java/org/sopt/certi/core/component/section/CertInfoSection.kt
