+++ app/src/main/java/org/sopt/certi/core/component/section/CertItemTitleSection.kt
