+++ app/src/main/java/org/sopt/certi/presentation/ui/mycertification/MyCertScreen.kt
@@ -1,25 +1,16 @@
 package org.sopt.certi.presentation.ui.mycertification
 
 import androidx.activity.compose.BackHandler
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
@@ -29,17 +20,15 @@ import kotlinx.collections.immutable.toImmutableList
 import org.sopt.certi.R
 import org.sopt.certi.core.component.dialog.CertiDeleteDialog
 import org.sopt.certi.core.component.header.MyPageHeader
-import org.sopt.certi.core.component.section.MyCertificationListItemSection
 import org.sopt.certi.core.state.UiState
-import org.sopt.certi.core.util.noRippleClickable
 import org.sopt.certi.core.util.screenWidthDp
 import org.sopt.certi.domain.model.certification.CertificationData
 import org.sopt.certi.presentation.type.MyCertType
-import org.sopt.certi.presentation.ui.mycertification.component.FavoriteCertItem
+import org.sopt.certi.presentation.ui.mycertification.component.CertificationList
+import org.sopt.certi.presentation.ui.mycertification.component.FavoriteCertList
 import org.sopt.certi.presentation.ui.mycertification.component.MyCertHeader
 import org.sopt.certi.presentation.ui.mycertification.state.MyCertUiState
 import org.sopt.certi.ui.theme.CERTITheme
-import org.sopt.certi.ui.theme.CertiTheme
 
 @Composable
 fun MyCertRoute(
@@ -66,6 +55,7 @@ fun MyCertRoute(
             certifications = state.data.toImmutableList(),
             onTabSelected = viewModel::updateSelectedTab,
             onEditModeToggle = viewModel::onEditModeToggle,
+            onCertificationClick = {}, // TODO: 자격증 연결 추가
             onFavoriteToggle = viewModel::onFavoriteToggle,
             onEditClick = viewModel::editItem,
             onDeleteClick = viewModel::openDeleteDialog,
@@ -84,6 +74,7 @@ fun CertificationScreen(
     certifications: ImmutableList<CertificationData>,
     onTabSelected: (MyCertType) -> Unit,
     onEditModeToggle: () -> Unit,
+    onCertificationClick: (Long) -> Unit,
     onFavoriteToggle: (Long) -> Unit,
     onEditClick: (Long) -> Unit,
     onDeleteClick: (Long) -> Unit,
@@ -104,116 +95,22 @@ fun CertificationScreen(
         }
 
         when (uiState.selectedTab) {
-            MyCertType.PLANNED -> {
-                CertList(
-                    isEditMode = uiState.isEditMode,
-                    certifications = certifications,
-                    onEditModeToggle = onEditModeToggle,
-                    onEditClick = onEditClick,
-                    onDeleteClick = onDeleteClick
-                )
-            }
-
-            MyCertType.ACQUIRED -> {
-                CertList(
-                    isEditMode = uiState.isEditMode,
-                    certifications = certifications,
-                    onEditModeToggle = onEditModeToggle,
-                    onEditClick = onEditClick,
-                    onDeleteClick = onDeleteClick
-                )
-            }
-
             MyCertType.FAVORITE -> {
                 FavoriteCertList(
                     certifications = certifications,
                     onFavoriteToggle = onFavoriteToggle
                 )
             }
-        }
-    }
-}
-
-@Composable
-private fun CertList(
-    isEditMode: Boolean,
-    certifications: ImmutableList<CertificationData>,
-    onEditModeToggle: () -> Unit,
-    onEditClick: (Long) -> Unit,
-    onDeleteClick: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val listState = rememberLazyListState()
-
-    LaunchedEffect(isEditMode) {
-        listState.scrollToItem(0)
-    }
-
-    LazyColumn(
-        modifier = modifier.fillMaxSize(),
-        state = listState,
-        contentPadding = PaddingValues(
-            start = screenWidthDp(20.dp),
-            top = screenWidthDp(16.dp),
-            end = screenWidthDp(20.dp),
-            bottom = screenWidthDp(20.dp)
-        ),
-        verticalArrangement = Arrangement.spacedBy(screenWidthDp(16.dp))
-    ) {
-        if (!isEditMode) {
-            item {
-                Text(
-                    text = stringResource(R.string.edit),
-                    style = CertiTheme.typography.body.semibold_16,
-                    color = CertiTheme.colors.gray400,
-                    textAlign = TextAlign.End,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .noRippleClickable(onEditModeToggle)
+            else -> {
+                CertificationList(
+                    isEditMode = uiState.isEditMode,
+                    certifications = certifications,
+                    onEditModeToggle = onEditModeToggle,
+                    onCertificationClick = onCertificationClick,
+                    onEditClick = onEditClick,
+                    onDeleteClick = onDeleteClick
                 )
             }
-        }
-
-        items(
-            items = certifications,
-            key = { it.certificationId }
-        ) { certification ->
-            MyCertificationListItemSection(
-                certificationData = certification,
-                isEditMode = isEditMode,
-                onCertificationClick = {},  // TODO: 자격증 이동 추가
-                onModifyClick = onEditClick,
-                onDeleteClick = onDeleteClick
-            )
-        }
-    }
-}
-
-@Composable
-private fun FavoriteCertList(
-    certifications: ImmutableList<CertificationData>,
-    onFavoriteToggle: (Long) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    LazyColumn(
-        modifier = modifier.fillMaxSize(),
-        contentPadding = PaddingValues(
-            start = screenWidthDp(20.dp),
-            top = screenWidthDp(24.dp),
-            end = screenWidthDp(20.dp),
-            bottom = screenWidthDp(20.dp)
-        ),
-        verticalArrangement = Arrangement.spacedBy(screenWidthDp(16.dp))
-    ) {
-        items(
-            items = certifications,
-            key = { it.certificationId }
-        ) { certification ->
-            FavoriteCertItem(
-                certificationData = certification,
-                onCertificationClick = {},  // TODO: 자격증 이동 추가
-                onFavoriteToggle = onFavoriteToggle
-            )
         }
     }
 }
@@ -239,6 +136,7 @@ private fun PreviewResumeMyCertScreen() {
             certifications = dummyCertifications.toImmutableList(),
             onTabSelected = { selectedTab = it },
             onEditModeToggle = {},
+            onCertificationClick = {},
             onFavoriteToggle = {},
             onEditClick = {},
             onDeleteClick = {}
